+++ src/js/engine/engine-loader.js
@@ -3,8 +3,10 @@
  * @since 3.3.0
  */
 
+import videojs from 'video.js';
+
 import RecordRTCEngine from './record-rtc';
-import {CONVERT_PLUGINS, TSEBML} from './convert-engine';
+import {CONVERT_PLUGINS, TSEBML, FFMPEGJS} from './convert-engine';
 import {RECORDRTC, LIBVORBISJS, RECORDERJS, LAMEJS, OPUSRECORDER, VMSG, WEBMWASM, GIFSHOT, AUDIO_PLUGINS} from './record-engine';
 
 /**
@@ -121,6 +123,11 @@ const getConvertEngine = function(convertEngine) {
         case TSEBML:
             // ts-ebml
             ConvertEngineClass = videojs.TsEBMLEngine;
+            break;
+
+        case FFMPEGJS:
+            // ffmpeg.js
+            ConvertEngineClass = videojs.FFmpegjsEngine;
             break;
 
         default: