+++ src/js/event.js
@@ -30,6 +30,7 @@ Event.ENUMERATE_READY = 'enumerateReady';
 Event.ENUMERATE_ERROR = 'enumerateError';
 Event.AUDIO_BUFFER_UPDATE = 'audioBufferUpdate';
 Event.AUDIO_OUTPUT_READY = 'audioOutputReady';
+Event.START_CONVERT = 'startConvert';
 Event.FINISH_CONVERT = 'finishConvert';
 Event.ENTER_PIP = 'enterPIP';
 Event.LEAVE_PIP = 'leavePIP';