+++ src/js/plugins/vmsg-plugin.js
@@ -3,6 +3,8 @@
  * @since 3.3.0
  */
 
+import videojs from 'video.js';
+
 import {Recorder} from 'vmsg';
 
 const RecordEngine = videojs.getComponent('RecordEngine');
@@ -14,6 +16,38 @@ const RecordEngine = videojs.getComponent('RecordEngine');
  * @augments videojs.RecordEngine
  */
 class VmsgEngine extends RecordEngine {
+    /**
+     * Creates an instance of this class.
+     *
+     * @param  {Player} player
+     *         The `Player` that this class should be attached to.
+     *
+     * @param  {Object} [options]
+     *         The key/value store of player options.
+     */
+    constructor(player, options) {
+        super(player, options);
+
+        /**
+         * Enables console logging for debugging purposes.
+         *
+         * @type {boolean}
+         */
+        this.debug = false;
+        /**
+         * Path to `vmsg.wasm` WebAssembly script.
+         *
+         * @type {string}
+         */
+        this.audioWebAssemblyURL = 'vmsg.wasm';
+        /**
+         * Additional configuration options for the vmsg library.
+         *
+         * @type {object}
+         */
+        this.pluginLibraryOptions = {};
+    }
+
     /**
      * Setup recording engine.
      *
@@ -28,10 +62,13 @@ class VmsgEngine extends RecordEngine {
         this.mediaType = mediaType;
         this.debug = debug;
 
+        // minimal default config
         this.config = {
             wasmURL: this.audioWebAssemblyURL
-            // XXX: support shimURL?
         };
+
+        // extend config with optional options
+        this.config = Object.assign(this.config, this.pluginLibraryOptions);
 
         this.engine = new Recorder(this.config,
             this.onRecordingAvailable.bind(this));