+++ test/defaults.spec.js
@@ -28,7 +28,9 @@ describe('pluginDefaultOptions', () => {
             debug: false,
             pip: false,
             autoMuteDevice: false,
+            videoBitRate: 1200,
             videoEngine: 'recordrtc',
+            videoFrameRate: 30,
             videoMimeType: 'video/webm',
             videoRecorderType: 'auto',
             videoWorkerURL: '',
@@ -46,7 +48,10 @@ describe('pluginDefaultOptions', () => {
             animationOptions: {},
             timeSlice: 0,
             convertEngine: '',
-            hotKeys: false
+            convertWorkerURL: '',
+            convertOptions: [],
+            hotKeys: false,
+            pluginLibraryOptions: {}
         });
     });
 });