+++ CHANGES.md
@@ -1,7 +1,39 @@
 videojs-record changelog
 ========================
 
-3.6.0 - unreleased
+3.8.0 - unreleased
+------------------
+
+- New ffmpeg.js plugin: convert recorded data into other audio/video file formats
+  in the browser (#201)
+- Support for specifying third-party plugin settings using the `pluginLibraryOptions`
+  option (#383)
+- New options: `videoBitRate` and `videoFrameRate` (currently only used in the
+  webm-wasm plugin)
+- New event: `startConvert` (used in ts-ebml and ffmpeg.js plugins) (#201)
+- Fix `RecordRTC.MediaStreamRecorder` import when using the `timeSlice` option
+- Examples: add `timeSlice` example demonstrating use of `timestamp` event
+- Bump required version for videojs-wavesurfer to 2.9.0 or newer for wavesurfer.js
+  3.0.0 support
+- Bump required version for recordrtc to 5.5.8
+- Bump required version for webrtc-adapter to 7.2.8 or newer
+
+
+3.7.1 - 2019/05/30
+------------------
+
+- Add missing video.js imports for plugins (#378)
+
+
+3.7.0 - 2019/05/27
+------------------
+
+- Add `setVideoInput(deviceId)` for changing the video input device (#369)
+- Add `setAudioInput(deviceId)` for changing the audio input device (#374)
+- Examples: use dark theme when possible
+
+
+3.6.0 - 2019/04/02
 ------------------
 
 - Add support for keyboard hotkeys (requires video.js 7.5.0 or newer) (#339)